update backend into MVC format & change ajax path as relative path
