+++ static/index.js
@@ -3,7 +3,7 @@ let nextPage=0;
 let keyword;
 let data=document.getElementById('image-data');
 function connect(url) {
-    url='http://127.0.0.1:3000/api/attractions?page=' + pageData.toString()
+    url='/api/attractions?page=' + pageData.toString()
     fetch(url, {
         method:'GET',
     })
@@ -68,7 +68,7 @@ let nextPage=1;
 const formData = new FormData(myForm);
 const queryKeyword = new URLSearchParams(formData);
 const params = Object.fromEntries(queryKeyword.entries());
-let queryUrl='http://127.0.0.1:3000/api/attractions?page=' + pageNumber.toString()+'&keyword='+params.keyword.toString();
+let queryUrl='/api/attractions?page=' + pageNumber.toString()+'&keyword='+params.keyword.toString();
 data.textContent="";
 observer.disconnect();
 fetch(queryUrl, {
